fix versioning
